requirements.txt
Django==4.2.7
django-allauth==0.57.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-extensions==3.2.3
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary>=2.9.10
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly==5.17.0
Pillow>=10.1.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
requests==2.31.0
openpyxl==3.1.2
xlrd==2.0.1
djangorestframework==3.16.1
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1
reportlab==4.4.5